+++ models/Producto.js
@@ -1,4 +1,5 @@
 import connection from "../utils/db.js";
+import { errorFalse, errorTrue } from "../middlewarm/manejoErrores.js";
 
 class Producto {
     constructor() {
@@ -29,21 +30,16 @@ class Producto {
         try {
             const [datos] = await this.getById(id)
             // consulto productos relacionados
-            
+            if (!datos) {
+                return errorTrue(id);
+            }
              // si no tiene se elimina 
             const [eliminado] = await connection.query(`delete from productos where id = ${id}`);
      
                 if (eliminado.affectedRows > 0) {
-                    return {
-                        error: false,
-                        message: `Se elimino con exito la categoria con id = ${id}`,
-                        data: datos
-                    }   
+                    return errorFalse(id, datos);
                 }else {
-                    return {
-                        error: true,
-                        message: `No se pudo eliminar ninguna producto con id = ${id}`
-                    }  
+                    return errorTrue(id);
                 }
             
         } catch (error) {
@@ -69,18 +65,10 @@ class Producto {
         try {
             const [result]  = await connection.query(query, params);
             if (result.affectedRows === 0) {
-                return {
-                    error: true,
-                    message: `No se pudo actualizar la productos con id = ${id}`,
-
-                }   
+                return errorTrue(id);
             }   
             
-            return {
-                error: false,
-                message: `Se Actualizo con exito la productis con id = ${id}`,
-                data :  result
-            }   
+            return errorFalse(id, result);
         } catch (error) {
             throw new Error("Error al Actualizar los productos");
         }
@@ -90,11 +78,6 @@ class Producto {
 
     async getById(id){
         const [consulta] = await connection.query(`select * from productos where id = ${id}`);  
-        
-        if (!consulta || consulta.length === 0) {
-            console.log('No se encontraron resultados para el id:', id);
-            return null;
-        }
         
         return consulta;
     }
